# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JobTracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  /** @nullable */
  name?: string | null;
  email: string;
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthUser {
  id: string;
  email: string;
  /** @nullable */
  name?: string | null;
  isAdmin: boolean;
  /** @nullable */
  emailVerifiedAt: string | null;
  /** True once the email-verification grace period has elapsed for an unverified account — the app must be fully blocked until verified. */
  mustVerify: boolean;
  /** ISO timestamp — when the grace period ends (createdAt + 48h) for an unverified account. */
  verifyGraceEndsAt: string;
}

export interface OtpRequestInput {
  email: string;
}

export interface OtpRequestResult {
  ok: boolean;
  /** Seconds to wait before requesting another code. */
  resendIn: number;
  alreadyVerified?: boolean;
}

export interface OtpLoginInput {
  email: string;
  code: string;
}

export interface PasswordResetInput {
  email: string;
  code: string;
  newPassword: string;
}

export interface VerifyConfirmInput {
  code: string;
}

export interface AuthProviders {
  google: boolean;
}

export interface AdminSetupStatus {
  adminExists: boolean;
}

export interface AdminSetupInput {
  token: string;
  email: string;
}

export interface AdminUser {
  id: string;
  email: string;
  /** @nullable */
  name?: string | null;
  isAdmin: boolean;
  /** @nullable */
  emailVerifiedAt: string | null;
  createdAt: string;
}

export interface StatusHistory {
  id: string;
  applicationId: string;
  /** @nullable */
  fromStatus?: string | null;
  toStatus: string;
  changedAt: string;
  /** @nullable */
  note?: string | null;
}

export interface Application {
  id: string;
  userId: string;
  company: string;
  role: string;
  /** @nullable */
  jobUrl?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  salary?: string | null;
  status: string;
  appliedDate: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
  statusHistory: StatusHistory[];
}

export interface ApplicationInput {
  company: string;
  role: string;
  /** @nullable */
  jobUrl?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  salary?: string | null;
  status?: string;
  /** @nullable */
  appliedDate?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface ApplicationUpdate {
  company?: string;
  role?: string;
  /** @nullable */
  jobUrl?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  salary?: string | null;
  status?: string;
  /** @nullable */
  appliedDate?: string | null;
  /** @nullable */
  notes?: string | null;
}

export type DashboardStatsByStatus = {[key: string]: number};

export interface DashboardStats {
  total: number;
  byStatus: DashboardStatsByStatus;
  stale: Application[];
  recent: Application[];
}

export type GoogleCallbackParams = {
code?: string;
state?: string;
};

export type ListApplicationsParams = {
search?: string;
status?: string;
sortBy?: string;
order?: string;
};