# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: JobTracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/setup:
    get:
      operationId: getAdminSetupStatus
      tags: [admin]
      summary: Whether an admin account already exists
      responses:
        "200":
          description: Setup status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSetupStatus"
    post:
      operationId: bootstrapAdmin
      tags: [admin]
      summary: Bootstrap the first admin account using a one-time setup token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminSetupInput"
      responses:
        "200":
          description: Admin created/promoted
        "400":
          description: Bad request
        "401":
          description: Invalid setup token
        "409":
          description: Admin account already exists
        "503":
          description: ADMIN_SETUP_TOKEN is not configured
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden — admin access required
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List applications
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
        - name: order
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
        "401":
          description: Unauthorized

    post:
      operationId: createApplication
      tags: [applications]
      summary: Create application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          description: Bad request
        "401":
          description: Unauthorized
  /applications/dashboard:
    get:
      operationId: getDashboard
      tags: [applications]
      summary: Get dashboard stats
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
        "401":
          description: Unauthorized
  /applications/export:
    get:
      operationId: exportApplications
      tags: [applications]
      summary: Export applications as CSV
      responses:
        "200":
          description: CSV file
        "401":
          description: Unauthorized
  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get application by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "401":
          description: Unauthorized
        "404":
          description: Not found

    patch:
      operationId: updateApplication
      tags: [applications]
      summary: Update application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "401":
          description: Unauthorized
        "404":
          description: Not found

    delete:
      operationId: deleteApplication
      tags: [applications]
      summary: Delete application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found
components:
  schemas:
...
    AdminSetupStatus:
      type: object
      required: [adminExists]
      properties:
        adminExists:
          type: boolean
    AdminSetupInput:
      type: object
      required: [token, email]
      properties:
        token:
          type: string
        email:
          type: string
    AdminUser:
      type: object
      required: [id, email, isAdmin, emailVerifiedAt, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: ["string", "null"]
        isAdmin:
          type: boolean
        emailVerifiedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    StatusHistory:
      type: object
      required: [id, applicationId, toStatus, changedAt]
      properties:
        id:
          type: string
        applicationId:
          type: string
        fromStatus:
          type: ["string", "null"]
        toStatus:
          type: string
        changedAt:
          type: string
        note:
          type: ["string", "null"]
    Application:
      type: object
      required: [id, userId, company, role, status, appliedDate, createdAt, updatedAt, statusHistory]
      properties:
        id:
          type: string
        userId:
          type: string
        company:
          type: string
        role:
          type: string
        jobUrl:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        salary:
          type: ["string", "null"]
        status:
          type: string
        appliedDate:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        statusHistory:
          type: array
          items:
            $ref: "#/components/schemas/StatusHistory"
    ApplicationInput:
      type: object
      required: [company, role]
      properties:
        company:
          type: string
        role:
          type: string
        jobUrl:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        salary:
          type: ["string", "null"]
        status:
          type: string
        appliedDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    ApplicationUpdate:
      type: object
      properties:
        company:
          type: string
        role:
          type: string
        jobUrl:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        salary:
          type: ["string", "null"]
        status:
          type: string
        appliedDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    DashboardStats:
      type: object
      required: [total, byStatus, stale, recent]
      properties:
        total:
          type: integer
        byStatus:
          type: object
          additionalProperties:
            type: integer
        stale:
          type: array
          items:
            $ref: "#/components/schemas/Application"
        recent:
          type: array
          items:
            $ref: "#/components/schemas/Application"