# lib/api-zod/src/generated/types/adminUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JobTracker API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminUser {
  id: string;
  email: string;
  /** @nullable */
  name?: string | null;
  isAdmin: boolean;
  /** @nullable */
  emailVerifiedAt: string | null;
  createdAt: string;
}